Added user Delete logic

+++ casino/src/main/java/com/tms/casino/controller/AdminController.java
@@ -4,12 +4,14 @@ import com.tms.casino.model.User;
 import com.tms.casino.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/admin")
+@RequestMapping("/admin")
 @RequiredArgsConstructor
 public class AdminController {
 
@@ -32,12 +34,18 @@ public class AdminController {
     }
 
     @DeleteMapping("/users/{userId}")
-    public ResponseEntity<String> deleteUser(@PathVariable Integer userId) {
+    public ResponseEntity<String> deleteUser(@PathVariable Integer userId, @AuthenticationPrincipal UserDetails userDetails) {
+
+        User currentUser = userService.getUserByUsername(userDetails.getUsername());
+        if (currentUser.getUserId().equals(userId)) {
+            return ResponseEntity.badRequest().body("Administrator cannot delete themselves.");
+        }
+
         try {
             userService.deleteUser(userId);
             return ResponseEntity.ok("User with id " + userId + " has been successfully deleted.");
         } catch (RuntimeException e) {
-            return ResponseEntity.status(404).body("User with id " + userId + " not found.");
+            return ResponseEntity.status(404).body("User with id " + userId + " was not found.");
         }
     }
 }
